Secured Package and Profile views

+++ packages/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from .models import Package
 from .forms import PackageForm
 
@@ -28,8 +29,13 @@ def package_detail(request, package_id):
     return render(request, 'packages/package_detail.html', context)
 
 
+@login_required
 def add_package(request):
     """ Add a package to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Admin can do this.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = PackageForm(request.POST, request.FILES)
         if form.is_valid():
@@ -49,8 +55,13 @@ def add_package(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_package(request, package_id):
     """ Edit a pacakge """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Admin can do this.')
+        return redirect(reverse('home'))
+
     package = get_object_or_404(Package, pk=package_id)
     if request.method == 'POST':
         form = PackageForm(request.POST, request.FILES, instance=package)
@@ -73,8 +84,13 @@ def edit_package(request, package_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_package(request, package_id):
     """ Delete a Package """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only Admin can do this.')
+        return redirect(reverse('home'))
+
     package = get_object_or_404(Package, pk=package_id)
     package.delete()
     messages.success(request, f'Package {package.name} deleted !')
